app/serializers/position.py
from rest_framework import serializers
from app.models import Position

class PositionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Position
        fields = ['id', 'title', 'rights_level']